requirements.txt
# Core Dependencies
langgraph>=0.0.20
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq
langchain-huggingface
langchain-community
# Configuration
pyyaml>=6.0

# Visualization
graphviz>=0.20

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0